Create html/css/js files for the project
Creates the html file that acts as the visual representation of
the project. In practice this means that it looks visually a lot
like the lichess.org since for the purposes of the project
we are acting like this would be a potential addition to lichess.
This commit creates the basic visual look of the header, which
will be fixed in future commits.