Choose which moves are significant
Adds the very simple way of defining which moves are potentially
significant when it comes to a pawn move being ''bad''. This is
done by looking at the combination of p-value and how often the
move appears in the lost and won games, percent-wise. Here, is
seems like there needs to be a constant which is used to multiply
the likelihood of having the move in lost games to allow wiggle
room that the very rough model needs.